+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - replaced the thirdparty/ folder with a git submodule in thirdparties/
 - now checking that a scope doesn't have our symbol before doing a `mut` operation (in dichotomic mode it was automatically handled, but not in linear mode)
 - enhancing the cmake defines (`-DARK_XYZ`) and the code using them
+- lighter Frame (from 40B to 32B), moved some unrelated logic from the frame to the virtual machine
 
 ### Removed